# task-1.1-2.1-2.3/backend/openapi.yaml
openapi: 3.0.0
info:
  title: Campaign Management API
  version: 1.0.0
  description: RESTful API for managing advertising campaigns with real-time notifications

servers:
  - url: http://localhost:3001
    description: Development server

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    Campaign:
      type: object
      properties:
        id:
          type: string
          format: uuid
        client_id:
          type: string
          format: uuid
        name:
          type: string
        status:
          type: string
          enum: [active, paused, completed, draft]
        budget:
          type: number
        spend:
          type: number
        impressions:
          type: integer
        clicks:
          type: integer
        conversions:
          type: integer
        ctr:
          type: number
        roas:
          type: number
        start_date:
          type: string
          format: date
        end_date:
          type: string
          format: date
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    Notification:
      type: object
      properties:
        id:
          type: string
          format: uuid
        campaign_id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        type:
          type: string
        title:
          type: string
        message:
          type: string
        is_read:
          type: boolean
        created_at:
          type: string
          format: date-time

    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string

paths:
  /auth/login:
    post:
      summary: Authenticate user
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  user:
                    type: object
        '401':
          description: Invalid credentials

  /campaigns:
    get:
      summary: List all campaigns
      tags: [Campaigns]
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: client_id
          schema:
            type: string
        - in: query
          name: sort
          schema:
            type: string
        - in: query
          name: order
          schema:
            type: string
            enum: [asc, desc]
      responses:
        '200':
          description: List of campaigns
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Campaign'
                  pagination:
                    type: object
                    properties:
                      page:
                        type: integer
                      limit:
                        type: integer
                      total:
                        type: integer
                      totalPages:
                        type: integer

    post:
      summary: Create a new campaign
      tags: [Campaigns]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [client_id, name, budget]
              properties:
                client_id:
                  type: string
                  format: uuid
                name:
                  type: string
                status:
                  type: string
                budget:
                  type: number
                start_date:
                  type: string
                  format: date
                end_date:
                  type: string
                  format: date
      responses:
        '201':
          description: Campaign created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Campaign'

  /campaigns/{id}:
    get:
      summary: Get campaign by ID
      tags: [Campaigns]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Campaign details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Campaign'
        '404':
          description: Campaign not found

    put:
      summary: Update campaign
      tags: [Campaigns]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                status:
                  type: string
                budget:
                  type: number
      responses:
        '200':
          description: Campaign updated
        '404':
          description: Campaign not found

    delete:
      summary: Soft delete campaign
      tags: [Campaigns]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Campaign deleted
        '404':
          description: Campaign not found

  /notifications:
    get:
      summary: Get user notifications
      tags: [Notifications]
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: unread_only
          schema:
            type: boolean
      responses:
        '200':
          description: List of notifications
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Notification'
                  unread_count:
                    type: integer